Add resources for X-Ray 1.5.03

+++ gamedata_cs/shaders/r3/common_cbuffers.h
@@ -4,9 +4,9 @@
 //	Used by dynamic lights and volumetric effects
 cbuffer	dynamic_light
 {
-	half4	Ldynamic_color;	// dynamic light color (rgb1)	- spot/point/sun
-	half4	Ldynamic_pos;	// dynamic light pos+1/range(w)	- spot/point
-	half4	Ldynamic_dir;	// dynamic light direction		- sun
+	float4	Ldynamic_color;	// dynamic light color (rgb1)	- spot/point/sun
+	float4	Ldynamic_pos;	// dynamic light pos+1/range(w)	- spot/point
+	float4	Ldynamic_dir;	// dynamic light direction		- sun
 }
 
 #endif	//	common_cbuffers_h_included
